Separate out the interactive cargo watch procedure
